Rename execute_context_base to stream_execute_context.rb
