+++ plugins/sharding/sources.am
@@ -11,4 +11,4 @@ sharding_scripts =				\
 	keys_parsable.rb			\
 	dynamic_columns.rb			\
 	window.rb			\
-	execute_context_base.rb
+	stream_execute_context.rb
